+++ CommandLine/src/Program.cs
@@ -3,6 +3,7 @@ using Autofac;
 using Autofac.Extensions.DependencyInjection;
 using McMaster.Extensions.CommandLineUtils;
 using Worms.Commands;
+using Worms.Container;
 
 namespace Worms
 {
